+++ karyon3-api/build.gradle
@@ -17,17 +17,7 @@
 apply plugin: 'java'
 
 dependencies {
-    compile project(":karyon3-api")
-    compile project(":karyon3-admin")
-    compile "com.netflix.governator:governator-core:${governator_version}"
-    
-    compile "com.google.inject:guice:${guice_version}"
-    compile "com.google.inject.extensions:guice-multibindings:${guice_version}"
-    compile "com.google.inject.extensions:guice-grapher:${guice_version}"
-
-    testCompile "org.hamcrest:hamcrest-core:${hamcrest_version}"
-    testCompile "org.hamcrest:hamcrest-library:${hamcrest_version}"
-    testCompile "org.mockito:mockito-all:1.9.5"
+    compile 'javax.inject:javax.inject:1'
 }
 
 eclipse {
