+++ karyon3-archaius2/src/main/resources/archaius2-local.properties
@@ -1,3 +1,5 @@
+# Default values for internal karyon environment properties
+
 karyon.domain=
 karyon.hostname=localhost
 karyon.publicHostname=localhost
@@ -14,6 +16,7 @@ karyon.asg=asg-dev
 karyon.cluster=cluster-dev
 karyon.stack=
 
+# Backwards compatible mappings for the legacy @ prefixed properties 
 @domain=${karyon.domain}
 @hostname=${karyon.hostname}
 @publicHostname=${karyon.publicHostname}
