+++ karyon3-archaius2/src/main/resources/archaius2.properties
@@ -1,0 +1,34 @@
+# This final contains the default mapping of Netflix EC2 environment settings
+# to Karyon's environment agnostic naming. 
+
+karyon.domain=${EC2_DOMAIN}
+karyon.hostname=${EC2_HOSTNAME}
+karyon.publicHostname=${EC2_PUBLIC_HOSTNAME}
+karyon.publicIpv4=${EC2_PUBLIC_IPV4}
+karyon.localHostname=${EC2_LOCAL_HOSTNAME}
+karyon.localIpv4=${EC2_LOCAL_IPV4}
+karyon.datacenter=cloud
+karyon.zoneId=${EC2_AVAILABILITY_ZONE}
+karyon.serverId=${EC2_INSTANCE_ID}
+karyon.ami=${EC2_AMI_ID}
+karyon.region=${EC2_REGION}
+
+# Backwards compatible mappings for the legacy @ prefixed properties 
+@domain=${karyon.domain}
+@hostname=${karyon.hostname}
+@publicHostname=${karyon.publicHostname}
+@publicIpv4=${karyon.publicIpv4}
+@localHostname=${karyon.localHostname}
+@localIpv4=${karyon.localIpv4}
+@datacenter=${karyon.datacenter}
+@region=${karyon.region}
+@zoneId=${karyon.zoneId}
+@cluster=${karyon.cluster}
+@ami=${karyon.ami}
+@asg=${karyon.asg}
+@serverId=${karyon.serverId}
+@stack=${karyon.stack}
+@environment=${karyon.environment}
+@appId=${karyon.appId}
+@countries=${karyon.countries}
+
