+++ karyon3-core/src/main/java/com/netflix/karyon/AbstractPropertySource.java
@@ -2,6 +2,8 @@ package com.netflix.karyon;
 
 import java.lang.reflect.Method;
 
+import com.netflix.karyon.spi.PropertySource;
+
 public abstract class AbstractPropertySource implements PropertySource {
     @Override
     public <T> T get(String key, Class<T> type) {
