+++ karyon3-core/src/main/java/com/netflix/karyon/PropertiesPropertySource.java
@@ -7,6 +7,7 @@ import javax.inject.Singleton;
 import com.google.inject.AbstractModule;
 import com.google.inject.Module;
 import com.google.inject.Provides;
+import com.netflix.karyon.spi.PropertySource;
 
 public class PropertiesPropertySource extends AbstractPropertySource {
     private Properties props;
