+++ karyon3-core/src/main/java/com/netflix/karyon/PropertySourceAutoBinder.java
@@ -2,6 +2,8 @@ package com.netflix.karyon;
 
 import com.google.inject.Binder;
 import com.google.inject.Key;
+import com.netflix.karyon.spi.AutoBinder;
+import com.netflix.karyon.spi.PropertySource;
 
 public class PropertySourceAutoBinder implements AutoBinder {
     @Override
