+++ karyon3-core/src/main/java/com/netflix/karyon/conditional/ConditionalOnPropertyMatcher.java
@@ -3,8 +3,8 @@ package com.netflix.karyon.conditional;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 
-import com.netflix.karyon.PropertySource;
 import com.netflix.karyon.conditional.annotations.ConditionalOnProperty;
+import com.netflix.karyon.spi.PropertySource;
 
 @Singleton
 public class ConditionalOnPropertyMatcher implements ConditionalMatcher<ConditionalOnProperty>{
