--- karyon3-core/src/main/java/com/netflix/karyon/experimental/admin/EnvAdminResource.java
@@ -1,15 +1,0 @@
-package com.netflix.karyon.experimental.admin;
-
-import javax.inject.Singleton;
-
-import com.netflix.karyon.admin.AdminService;
-
-import java.util.Map;
-
-@Singleton
-@AdminService(name="env", index="list")
-final class EnvAdminResource {
-    public Map<String, String> list() throws Exception {
-        return System.getenv();
-    }
-}
